[Rust] Code movement of Player and Commands
Move code into separate files, partly just a learning exercise on how to keep
things neat.
